app/requirements.txt
Pillow==10.1.0
pyyaml==5.3.1
psycopg2-binary==2.9.6
django==4.2.1
djangorestframework==3.14.0
django-filter==23.2
django-rest-passwordreset==1.3.0
django-phonenumber-field[phonenumbers]==7.1.0
django-imagekit==5.0.0
django-cors-headers==4.2.0
django-baton==2.8.0
drf-spectacular[sidecar]==0.26.5
social-auth-app-django==5.4.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
model-bakery==1.15.0
pytest==7.4.2
pytest-django==4.5.2
pytest-cov==4.1.0